Set eventstream task to longrunning
